Final working version
The script described is designed to generate machine learning inputs based on feature space derived from either 1H or 13C NMR spectra and to model a given property. The input file is a CSV file containing the chemical compound name, its SMILES code, and the label corresponding to the property to be modeled.